+++ styles/easy_refresh_halloween/lib/src/halloween_indicator.dart
@@ -23,28 +23,31 @@ class _HalloweenIndicator extends StatefulWidget {
 }
 
 class _HalloweenIndicatorState extends State<_HalloweenIndicator> {
+  RuntimeArtboard? _artboard;
   late SimpleAnimation _pullController;
   late SimpleAnimation _triggerController;
   late SimpleAnimation _loadingController;
 
   double get _offset => widget.state.offset;
 
-  bool _reset = true;
-
   IndicatorMode get _mode => widget.state.mode;
+
+  double get _actualTriggerOffset => widget.state.actualTriggerOffset;
 
   @override
   void initState() {
     super.initState();
-    _pullController = SimpleAnimation(
+    _pullController = OneShotAnimation(
       'Pull',
-      autoplay: true,
+      autoplay: false,
     );
     _triggerController = _OneShotCustomAnimation(
       'Trigger',
       autoplay: false,
       onStop: () {
-        _loadingController.isActive = true;
+        if (_mode == IndicatorMode.processing) {
+          _loadingController.isActive = true;
+        }
       },
     );
     _loadingController = SimpleAnimation(
@@ -65,59 +68,66 @@ class _HalloweenIndicatorState extends State<_HalloweenIndicator> {
 
   /// Mode change listener.
   void _onModeChange(IndicatorMode mode, double offset) {
-    if (mode == IndicatorMode.inactive) {
-      _reset = true;
-      if (_loadingController.isActive) {
-        _loadingController.isActive = false;
+    if (mode == IndicatorMode.ready) {
+      if (!_triggerController.isActive) {
+        _triggerController.isActive = true;
+      }
+      return;
+    }
+    if (mode == IndicatorMode.drag || mode == IndicatorMode.armed) {
+      if (_triggerController.isActive) {
+        _triggerController.isActive = false;
       }
       _triggerController.reset();
-      _loadingController.reset();
-    } else {
-      if (_reset) {
-        _reset = false;
-        _pullController.isActive = true;
-      }
+      _triggerController.instance?.animation.apply(0, coreContext: _artboard!);
     }
-    if (mode == IndicatorMode.ready) {
-      _triggerController.isActive = true;
+    if (mode == IndicatorMode.inactive) {
+      _loadingController.isActive = false;
+      _loadingController.reset();
+      _loadingController.instance?.animation.apply(0, coreContext: _artboard!);
+      return;
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    if (_artboard != null) {
+      if (_mode == IndicatorMode.drag || _mode == IndicatorMode.armed) {
+        final scale = (_offset / _actualTriggerOffset).clamp(0.0, 1.0);
+        _pullController.instance?.animation
+            .apply(scale, coreContext: _artboard!);
+      }
+    }
     return SizedBox(
       width: double.infinity,
       height: _offset,
-      child: _offset == 0 && _mode == IndicatorMode.inactive
-          ? const SizedBox()
-          : RiveAnimation.asset(
-              'packages/easy_refresh_halloween/assets/halloween.riv',
-              controllers: [
-                _pullController,
-                _triggerController,
-                _loadingController,
-              ],
-              fit: BoxFit.cover,
-            ),
+      child: RiveAnimation.asset(
+        'packages/easy_refresh_halloween/assets/halloween.riv',
+        controllers: [
+          _pullController,
+          _triggerController,
+          _loadingController,
+        ],
+        fit: BoxFit.cover,
+        onInit: (artboard) {
+          if (artboard is RuntimeArtboard) {
+            _artboard = artboard;
+          }
+        },
+      ),
     );
   }
 }
 
-T? _ambiguate<T>(T? value) => value;
-
 class _OneShotCustomAnimation extends SimpleAnimation {
   /// Fires when the animation stops being active
   final VoidCallback? onStop;
-
-  /// Fires when the animation starts being active
-  final VoidCallback? onStart;
 
   _OneShotCustomAnimation(
     String animationName, {
     double mix = 1,
     bool autoplay = true,
     this.onStop,
-    this.onStart,
   }) : super(animationName, mix: mix, autoplay: autoplay) {
     isActiveChanged.addListener(onActiveChanged);
   }
@@ -132,10 +142,9 @@ class _OneShotCustomAnimation extends SimpleAnimation {
   /// Perform tasks when the animation's active state changes
   void onActiveChanged() {
     // Fire any callbacks
-    isActive
-        ? onStart?.call()
-        // onStop can fire while widgets are still drawing
-        : _ambiguate(WidgetsBinding.instance)
-            ?.addPostFrameCallback((_) => onStop?.call());
+    if (!isActive) {
+      reset();
+      onStop?.call();
+    }
   }
 }