+++ styles/easy_refresh_space/lib/easy_refresh_space.dart
@@ -1,11 +1,9 @@
 library easy_refresh_space;
 
-import 'package:flare_flutter/flare.dart';
-import 'package:flare_flutter/flare_actor.dart';
-import 'package:flare_flutter/flare_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/physics.dart';
 import 'package:easy_refresh/easy_refresh.dart';
+import 'package:rive/rive.dart';
 
 part 'src/space_indicator.dart';
 part 'src/space_header.dart';