+++ styles/easy_refresh_halloween/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+> perf: Optimize animation loading.  
+> fix: The animation does not stop when [IndicatorMode.ready] returns to [IndicatorMode.armed].  
+
 ## 1.0.0+1
 > fix: dart >=2.13.0.