+++ styles/easy_refresh_skating/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.1
+> perf: Optimize animation loading.  
+
 ## 1.0.0+1
 > fix: dart >=2.13.0.